+++ zvtm/trunk/src/main/java/fr/inria/zvtm/engine/OffscreenViewPanel.java
@@ -81,12 +81,12 @@ public class OffscreenViewPanel extends ViewPanel {
 		synchronized(drawnGlyphs){
 		    drawnGlyphs.removeAllElements();
 		    float uncoef=(float)((cams[nbcam].focal+cams[nbcam].altitude)/cams[nbcam].focal);
-		    long viewW=this.getSize().width;//compute region's width and height
-		    long viewH=this.getSize().height;
-		    long viewEC = (long)(cams[nbcam].posx+viewW/2*uncoef);
-		    long viewNC = (long)(cams[nbcam].posy+viewH/2*uncoef);
-		    long viewWC = (long)(cams[nbcam].posx-viewW/2*uncoef);
-		    long viewSC = (long)(cams[nbcam].posy-viewH/2*uncoef);
+		    double viewW = this.getSize().width;//compute region's width and height
+		    double viewH = this.getSize().height;
+		    double viewEC = cams[nbcam].posx + viewW/2*uncoef;
+		    double viewNC = cams[nbcam].posy + viewH/2*uncoef;
+		    double viewWC = cams[nbcam].posx - viewW/2*uncoef;
+		    double viewSC = cams[nbcam].posy - viewH/2*uncoef;
 		    gll = cams[nbcam].parentSpace.getDrawingList();
 		    for (int i=0;i<gll.length;i++){
 			if (gll[i].visibleInViewport(viewWC, viewNC, viewEC, viewSC, cams[nbcam])){
